Move the TileDBArrayWrapper for xarray into its own file and clean-up
